ninjareact/requirements.txt
django
django-ninja
djangorestframework
uvicorn